+++ backend/tastify_backend/urls.py
@@ -2,13 +2,13 @@ from django.contrib import admin
 from django.urls import include, path
 from django.conf import settings
 from django.conf.urls.static import static
+from core.views import health
 
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('api/users/', include('apps.users.urls')),
-    path('api/', include('core.urls')),
-    path('api/', include('apps.menu.urls')),
-    path('api/', include('apps.tables.urls')),
+    path('api/health/', health, name='health'),
+    path('api/', include('tastify_backend.api_router')),
 ]
 
 if settings.DEBUG: